+++ processing/src/feat_comp.cc
@@ -24,7 +24,15 @@ std::map<std::string, float> FeatComp::process(const LorentzVector& b_1,
                                                Year year,
                                                const int& n_vbf,
                                                const bool& svfit_conv,
-                                               const bool& hh_kinfit_conv) {
+                                               const bool& hh_kinfit_conv,
+                                               const float& b_1_cvsl,
+										       const float& b_2_cvsl,
+										       const float& vbf_1_cvsl,
+										       const float& vbf_2_cvsl,
+										       const float& b_1_cvsb,
+										       const float& b_2_cvsb,
+										       const float& vbf_1_cvsb,
+										       const float& vbf_2_cvsb) {
     /* Compute HL features from base event*/
 
     bool use_vbf = n_vbf >= 2;
@@ -46,7 +54,15 @@ std::map<std::string, float> FeatComp::process(const LorentzVector& b_1,
     if (FeatComp::_feat_check("year"))       feats["year"]       = year;
     if (FeatComp::_feat_check("svfit_conv")) feats["svfit_conv"] = svfit_conv;
     if (FeatComp::_feat_check("n_vbf"))      feats["n_vbf"]      = n_vbf;
-    FeatComp::_add_jet_flags(b_1_csv, b_2_csv, feats);
+    if (EvtProc::_feat_check("b_1_cvsl"))    feats["b_1_cvsl"]   = FeatComp::_get_cvsb_flag(year, b_1_cvsl);
+    if (EvtProc::_feat_check("b_2_cvsl"))    feats["b_2_cvsl"]   = FeatComp::_get_cvsb_flag(year, b_2_cvsl);
+    if (EvtProc::_feat_check("vbf_1_cvsl"))  feats["vbf_1_cvsl"] = use_vbf ? FeatComp::_get_cvsb_flag(year, vbf_1_cvsl) : std::nanf("1");
+    if (EvtProc::_feat_check("vbf_2_cvsl"))  feats["vbf_2_cvsl"] = use_vbf ? FeatComp::_get_cvsb_flag(year, vbf_2_cvsl) : std::nanf("1");
+    if (EvtProc::_feat_check("b_1_cvsb"))    feats["b_1_cvsb"]   = FeatComp::_get_cvsb_flag(year, b_1_cvsb);
+    if (EvtProc::_feat_check("b_2_cvsb"))    feats["b_2_cvsb"]   = FeatComp::_get_cvsb_flag(year, b_2_cvsb);
+    if (EvtProc::_feat_check("vbf_1_cvsb"))  feats["vbf_1_cvsb"] = use_vbf ? FeatComp::_get_cvsb_flag(year, vbf_1_cvsb) : std::nanf("1");
+    if (EvtProc::_feat_check("vbf_2_cvsb"))  feats["vbf_2_cvsb"] = use_vbf ? FeatComp::_get_cvsb_flag(year, vbf_2_cvsb) : std::nanf("1");
+    FeatComp::_add_btag_flags(year, b_1_csv, b_2_csv, feats);
 
     // Delta phi
     if (FeatComp::_feat_check("dphi_l1_l2"))      feats["dphi_l1_l2"]      = FeatComp::delta_phi(l_1, l_2);
@@ -239,14 +255,26 @@ inline float FeatComp::calc_cos_delta(const LorentzVector& v, const LorentzVecto
     return  CosTheta(boost(v, r.BoostToCM()), r);
 }
 
-void FeatComp::_add_jet_flags(const float& b_1_csv, const float& b_2_csv, std::map<std::string, float>& feats) {
+void FeatComp::_add_btag_flags(Year year, const float& b_1_csv, const float& b_2_csv, std::map<std::string, float>& feats) {
     int tag_1(0), tag_2(0);
-    for (float wp : (_use_deep_bjet_wps ? _deep_bjet_wps : _bjet_wps)) {
+    for (float wp : (_use_deep_bjet_wps ? _deep_bjet_wps : _bjet_wps)[year]) {
         if (b_1_csv >= wp) tag_1++;
         if (b_2_csv >= wp) tag_2++;
     }
     if (FeatComp::_feat_check("jet_1_quality")) feats["jet_1_quality"] = tag_1;
     if (FeatComp::_feat_check("jet_2_quality")) feats["jet_2_quality"] = tag_2;
+}
+
+int FeatComp::_get_cvsl_flag(Year year, const float& score) {
+    int tag(0);
+    for (float wp : _cvsl_wps[year]) if (score >= wp) tag++;
+    return tag;
+}
+
+int FeatComp::_get_cvsb_flag(Year year, const float& score) {
+    int tag(0);
+    for (float wp : _cvsb_wps[year]) if (score <= wp) tag++;
+    return tag;
 }
 
 inline float FeatComp::calc_centrality( const LorentzVector& v, const LorentzVector& VBFjet1, const LorentzVector& VBFjet2)